uterm_vt: split the file into the fake and real backend
So it would be easier to add a new backend using libseat
