+++ src/meson.build
@@ -100,6 +100,8 @@ uterm_srcs = [
   'uterm_video.c',
   'uterm_monitor.c',
   'uterm_vt.c',
+  'uterm_vt_fake.c',
+  'uterm_vt_real.c',
   'uterm_input.c',
   'uterm_input_uxkb.c',
   'uterm_input_pointer.c',
